requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile .\pyproject.toml -o requirements.txt
aiofiles==25.1.0
    # via aiogram
aiogram==3.24.0
    # via caloriescounter (./pyproject.toml)
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via aiogram
aiosignal==1.4.0
    # via aiohttp
aiosqlite==0.22.1
    # via caloriescounter (./pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
attrs==25.4.0
    # via aiohttp
certifi==2026.1.4
    # via
    #   aiogram
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
distro==1.9.0
    # via google-genai
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
google-auth==2.47.0
    # via google-genai
google-genai==1.56.0
    # via caloriescounter (./pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
magic-filter==1.0.12
    # via aiogram
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via
    #   aiogram
    #   google-genai
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.2.1
    # via caloriescounter (./pyproject.toml)
requests==2.32.5
    # via
    #   google-auth
    #   google-genai
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via google-genai
tenacity==9.1.2
    # via google-genai
typing-extensions==4.15.0
    # via
    #   aiogram
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.2
    # via requests
websockets==15.0.1
    # via google-genai
yarl==1.22.0
    # via aiohttp
